requirements.txt
setuptools==45.0.0
python-decouple
Django==5.0.4
djangorestframework==3.15.2
drf-yasg==1.21.7
gunicorn
whitenoise
python-dotenv

